qt layer widget own file
